Add --help support to utilities, unit test that utils support --help
This adds a unit test which requires that all utilities in the "bin"
directory must respond to -h and --help flags with something that looks
like usage information. There's a list of exceptions, which currently
includes all of the Javascript because the Javascript scripts won't
actually run from a unit test environment (they can't find stuff for
require() unless they're in their installed location).

Some utilities got new documentation text; some had their documentation
text reformatted to look like usage information to the unit test. Two
Python scripts are modified to import a dependency after parsing their
command-line args, instead of before.

This adds a unit test which depends on nodejs being version 0.12 or
later, where previously nodejs was supported as far back as 0.10. Other
branches have adopted newer versions of nodejs, and made the appropriate
setup-script changes, so this should be merged only after they are. If
this does end up merged to master, it will produce a failing unit test
but won't break actual execution.

Test plan: The newly-added unit test confirms that all of the modified
scripts are still at least able to run. The changes are minimal, so
subtler breakage is unlikely. Also install and run on my own rig, with
an up-to-date nodejs.