Add a file for commonizing bash functions
Currently there is a fair amount of redundancy between the shell scripts
in oref0, because if a function could be shared between scripts but
isn't part of the public interface and isn't worth the overhead of its
own separate script, there wasn't anywhere to put it. This introduces a
file oref0-bash-common-functions.sh, for stuff that's shared between
scripts. This commit moves only one function there, "overtemp", which was
previously copy-pasted into three places.

Test plan: Install and run on my own rig. Moderate risk of conspicuous
breakage, negligible risk of subtle breakage.