Add more unit-test infrastructure
This adds two unit tests. The first goes through the source and attempts
to compile/parse every shell, python3, and Javascript file. This will
catch most syntax errors. The second looks for shell-scripts in the test
directory, and runs them as unit tests. This will be used to define unit
tests written in shell, of shell functions.

Because Python is checked as Python 3, some Python 2 files were fixed to
also be Python 3 (though they'll still be run as Python 2, for now), by
changing print statements to have parentheses, and (in a file where they
were mixed), converting tabs to spaces.