Add script for sending specified Pushover notification
Add oref0-send-notification.sh. In later commits, this will be used for
a variety of Pushover notifications, extending the currently available
set (which contains only one notification, carbs-required). The script
handles cooldowns and configurability, in a way that makes it easy to
get a cooldown and a full set of config options for each alert without
any per-alert extra work.

This introduces an upgrade script, oref0-upgrade-version.sh. This is
needed because the setup script gets the user's Pushover API
information, and stores it as arguments to oref0-pushover.sh - but
doesn't store it in preferences.json or anywhere accessible from outside
of oref0-pushover.sh. Since part of the goal is to make more Pushover
notifications, sent from more places, we need to migrade the Pushover
API information to somewhere more accessible, namely, preferences.json;
oref0-upgrade-version.sh does that migration. The upgrade script is
idempotent, and it is intended that it will get run whenever users
upgrade to a new oref0 version (but this is not currently actually done).

Test plan: Small components of this are unit-tested. For
notification-sending, I'm going to install this and manually send some
notifications and see if they arrive and look correct (since mocking
curl and a Pushover-like API endpoint would be more complicated and error
prone than just testing by hand).