Refactor crontab creation in oref0-setup.sh
Currently OpenAPS uses a large number of cronjobs, created by
oref0-setup.sh.  I plan to consolidate these into a smaller number of
cronjobs which do more things, and which keep their settings in the main
config file rather than in the crontab. As a first step, this commit
refactors oref0-setup.sh to make the cronjob creation more legible;
previously each one used an idiom to check whether it was already there
and create it if it isn't, that idiom is now part of the add_to_crontab
function.

There is some risk of breakage due to subtleties of bash quoting here,
and some branches correspond to hardware I don't have, so this is not
fully tested. What I have done is proofread the diff very thoroughly.
Most of these are going to be refactored into new, less-device-specific
shell scripts in the next commit, which will moot those issues if there
are any.