Refactor cronjobs into a more legible set of scripts
There are a bunch of things in oref0 that happen at regular intervals,
like syncing with a pump and CGM, checking for low battery conditions,
etc. These are handled by cron. Previously, oref0-setup.sh would upsert
a large set of cronjobs and put them in crontab, and these would be
slightly different based on aspects of your setup (eg which model CGM
and pump). But the way this was done meant that these commands had to be
squashed onto one line, surrounded by an extra level of quoting, and
couldn't use any utility functions, which made it very hard to read and
understand what was going on. This patch replaces that with a set of
scripts, oref0-cron-*.sh, run from cron, which consolidate all the
complexity that was previously living in the crontab itself.

Various settings that were used by these cron tasks now live in
preferences.json; oref0-setup.sh creates them for new installs, and
oref0-runagain.sh adds them when upgrading. oref0-upgrade-version.sh
removes the obsolete cron tasks; currently it needs to be run manually
on upgrade in addition to oref0-runagain.sh, but a later commit will
address that.

Note: Significant portions of this are hardware-specific in a way that
means I can't test them myself, and there are no preexisting unit tests.
I've done my best to do this refactor in a reliably-behavior-preserving
way, but I can't test hardware that I don't have; so this is relying on
there being at least one developer-user for each hardware configuration
in between the "dev" branch and general releases.