Add notifications for long uptime, IP address change
Add two Pushover notifications: IP address changed, showing what the new IP
address is so it's easier to figure out where to SSH to; and long
uptime, so that when using a USB battery with no software-accessible
battery indicator, there can still be some sort of reminder to change
batteries. Both notifications are disabled by default, and this also
adds settings for enabling/disabling notification types.