Unit test to syntax-check script and source files
This adds a unit test which goes through the source (everything in
oref0/bin and oref0/lib) and attempts to compile/parse each Python,
Javascript and shell file, failing the test if there are syntax errors.
For Python, this is done with Python 3, regardless of whether the file
specifies Python 2 or Python 3 (currently there are some files using
each). Since Python 3 is stricter, some files were changed to pass
Python 3's stricter syntax rules, in particular, converting mixed spaces
and tabs to all spaces, and putting parentheses on print() statements.