Refactor: Refer to scripts by their names rather than indirectly
There is an "oref0" script, which, if given a parameter it doesn't
recognize like "foo", will find and run "oref0-foo". This makes it
tricky to reliably find all invocations of a script. Remove usages of
this implied-oref0-prefix hack, in preparation for getting rid of the
oref0 script entirely.