Convert tempBasalFunctions argument to a regular require
determine_basal took an argument tempBasalFunctions, which was always
the basal-set-temp.js module. In principle that could've been used to
swap in some other module, but this was never used and doesn't seem very
useful given what's actually in that module. Remove that argument, make
it a regular direct require(), and adjust calls to not pass the removed
argument.