Switch determine-basal from using error-objects to exceptions
Before, determine-basal would return an object like
{error:"Errormessage"} if it encountered a problem. Switch that to
throwing exceptions (and change its callers accordingly).

This revealed one bug in a unit test, where a profile had "max_bg " in
it (with an extra space) instead of "max_bg", and this wasn't getting
caught. The switch to exceptions made that test fail (as it should
have). Fixed that test by removing the stray space.