Refactor: Don't depend on grepping the "reason" field for carbsReq notifications
The "reason" field in enact/smb-suggested.json is primarily a
monitoring/debugging mechanism: a human-readable string alongside a
temp-basal or SMB decision to explain why it did or didn't do something.
But there's also a hack where oref0-pushover looks at it (technically,
looks at the whole file) for the strings "add'l" and "maxBolus" to
decide whether it needs to send a Pushover alert telling the user they
need to eat carbs.

Add a separate field notifyUser to contain that, instead. Not
refactoring oref0-pushover.sh just yet, just severing its dependency on
the "reason" field.